Strategy merged and javadoc updated
