+++ src/designProject/KeyboardInput.java
@@ -9,8 +9,8 @@ class KeyboardInput extends KeyAdapter implements KeyListener {
 	Speed fastKnight = new FastSpeedDecorator((Speed) new Knight()); 
 	int fastSpeed = fastKnight.get_velocity();
 	
-	//Speed extraFastRunner = new ExtraFastSpeedDecorator((Speed) new Runner()); 
-	//int extraFastSpeed = extraFastRunner.get_velocity();
+	Speed extraFastRunner = new ExtraFastSpeedDecorator((Speed) new Runner());
+	int extraFastSpeed = extraFastRunner.get_velocity();
 	
 	Speed slowSpeedDefender = new SlowSpeedDecorator((Speed) new Defender());
 	int slowSpeed = slowSpeedDefender.get_velocity();
@@ -34,28 +34,28 @@ class KeyboardInput extends KeyAdapter implements KeyListener {
 	
 		//Defender
 		if ((e.getKeyCode()== KeyEvent.VK_W)) {
-			//Game.changeY(-1);
+			Game.changeY(-1);
 			slowSpeedDefender.changeY(-1, slowSpeed);
 			}
 			
 		else if (e.getKeyCode()== KeyEvent.VK_A) {
-			//Game.changeX(-1);
+			Game.changeX(-1);
 			slowSpeedDefender.changeX(-1,slowSpeed);						
 			}
 				
 		else if (e.getKeyCode()== KeyEvent.VK_S) {
-			//Game.changeY(1);
+			Game.changeY(1);
 			slowSpeedDefender.changeY(1,slowSpeed);
 			}
 				
 		else if (e.getKeyCode()== KeyEvent.VK_D) {
-			//Game.changeX(1);
+			Game.changeX(1);
 			slowSpeedDefender.changeX(1,slowSpeed);
 			}
 		
 		//Knight
 		else if (e.getKeyCode()== KeyEvent.VK_UP) {
-			//Game.changeY(-1);
+			Game.changeY(-1);
 			fastKnight.changeY(-1, fastSpeed);
 			}
 			
@@ -82,7 +82,15 @@ class KeyboardInput extends KeyAdapter implements KeyListener {
 		else if (e.getKeyCode()== KeyEvent.VK_DELETE) {
              Originator.loadState();
 			}
-		
+
+		else if (e.getKeyCode() == KeyEvent.VK_O)
+		{
+			Game.context = new StrategyContext(new StrategyEasy());
+		}
+		else if (e.getKeyCode() == KeyEvent.VK_P)
+		{
+			Game.context = new StrategyContext(new StrategyHard());
+		}
 		}
 
 	public void keyTyped(KeyEvent e) {}
