+++ src/designProject/Strategy.java
@@ -1,3 +1,8 @@
+package designProject;
+
+/**
+ * Strategy Pattern interface.
+ */
 public interface Strategy {
 	public void damage();
 	public void Attack();
