--- strategyctf/src/IMapFactory.java
@@ -1,4 +1,0 @@
-
-public interface IMapFactory{
-	public MapObject getObject(String objectType);
-}
