--- strategyctf/src/KeyboardInput.java
@@ -1,79 +1,0 @@
-
-
-import java.awt.Graphics;
-import java.awt.event.ActionEvent;
-import java.awt.event.KeyAdapter;
-import java.awt.event.KeyEvent;
-import java.util.ArrayList;
-
-class KeyboardInput extends KeyAdapter  {
-	
-	Speed fastKnight = new FastSpeedDecorator((Speed) new Knight()); 
-	int fastSpeed = fastKnight.get_velocity();
-	
-	Speed extraFastRunner = new ExtraFastSpeedDecorator((Speed) new Runner()); 
-	int extraFastSpeed = extraFastRunner.get_velocity();
-	
-	Speed slowSpeedDefender = new SlowSpeedDecorator((Speed) new Defender());
-	int slowSpeed = slowSpeedDefender.get_velocity();
-	
-	
-	
-	public void keyPressed( KeyEvent e) {
-		
-		
-		if (e.getKeyCode()== KeyEvent.VK_W) {
-			Game.changeY(-1);
-			fastKnight.changeY(-1,fastSpeed);
-			extraFastRunner.changeY(-1,extraFastSpeed);
-			slowSpeedDefender.changeY(-1, slowSpeed);
-			}
-		
-		else if (e.getKeyCode()== KeyEvent.VK_A) {
-			Game.changeX(-1);
-			fastKnight.changeX(-1,fastSpeed);
-            extraFastRunner.changeX(-1,extraFastSpeed);
-            slowSpeedDefender.changeX(-1, slowSpeed);
-			}
-		
-		else if (e.getKeyCode()== KeyEvent.VK_S) {
-			Game.changeY(1);
-			fastKnight.changeY(1,fastSpeed);
-			extraFastRunner.changeY(1,extraFastSpeed);
-			slowSpeedDefender.changeY(1,slowSpeed);
-			}
-		
-		else if (e.getKeyCode()== KeyEvent.VK_D) {
-			Game.changeX(1);
-			fastKnight.changeX(1,fastSpeed);
-            extraFastRunner.changeX(1, extraFastSpeed);
-            slowSpeedDefender.changeX(1,slowSpeed);
-			}
-		
-		 else if (e.getKeyCode()== KeyEvent.VK_CONTROL){
-             Originator.saveState();
-             
-		 	}
-		
-		 else if (e.getKeyCode()== KeyEvent.VK_DELETE) {
-             Originator.loadState();
-		 }
-		
-		if (e.getKeyCode() == KeyEvent.VK_O)
-		{
-			Game.context = new StrategyContext(new StrategyEasy());
-		}
-		else if (e.getKeyCode() == KeyEvent.VK_P)
-		{
-			Game.context = new StrategyContext(new StrategyHard());
-		}
-	}
-	
-	public void keyTyped(KeyEvent e) {}
-	public void keyReleased(KeyEvent e) {}
-
-	
-
-	
-   
-}
