--- strategyctf/src/SpeedDecorator.java
@@ -1,36 +1,0 @@
-
-import java.awt.Graphics;
-import java.awt.Graphics2D;
-
-public abstract class SpeedDecorator implements Speed {
-	
-	protected  Speed decoratedSpeed;
-
-	public SpeedDecorator( Speed decoratedSpeed ){
-		this.decoratedSpeed = decoratedSpeed;
-	}
-	
-	@Override
-	public int get_velocity() {
-		return decoratedSpeed.get_velocity();
-		}
-	
-	@Override
-	public void draw_character(Graphics g) {
-		decoratedSpeed.draw_character(g);
-	}
-	
-	@Override
-	public void changeX(int offset, int velocity) {
-		decoratedSpeed.changeX(offset,velocity);
-	}
-
-	@Override
-	public void changeY(int offset, int velocity ) {
-		decoratedSpeed.changeY(offset, velocity);
-		
-	}
-	
-	
-
-}
